Move inspector classes to correct package
