+++ kt/plugins/godot-intellij-plugin/src/main/kotlin/godot/intellij/plugin/inspections/registration/inspector/RegisterSignalInspection.kt
@@ -1,33 +1,17 @@
-package godot.intellij.plugin.inspections.registration.property.inspector
+package godot.intellij.plugin.inspections.registration.inspector
 
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.psi.PsiElementVisitor
 import godot.annotation.RegisterSignal
-import godot.intellij.plugin.inspections.registration.property.quickfix.ClassNotRegisteredQuickFix
-import godot.intellij.plugin.inspections.registration.property.quickfix.RegisterPropertyMutabilityQuickFix
-import godot.intellij.plugin.inspections.registration.property.quickfix.RegisterSignalInitializerQuickFix
-import godot.intellij.plugin.inspections.registration.property.quickfix.RegisterSignalMutabilityQuickFix
-import org.jetbrains.kotlin.backend.common.serialization.findPackage
-import org.jetbrains.kotlin.builtins.KotlinBuiltIns
-import org.jetbrains.kotlin.descriptors.impl.ClassConstructorDescriptorImpl
-import org.jetbrains.kotlin.descriptors.impl.SimpleFunctionDescriptorImpl
-import org.jetbrains.kotlin.idea.caches.resolve.analyze
+import godot.intellij.plugin.inspections.registration.quickfix.RegisterSignalInitializerQuickFix
+import godot.intellij.plugin.inspections.registration.quickfix.RegisterSignalMutabilityQuickFix
 import org.jetbrains.kotlin.idea.inspections.AbstractKotlinInspection
-import org.jetbrains.kotlin.idea.intentions.loopToCallChain.isConstant
 import org.jetbrains.kotlin.idea.util.findAnnotation
 import org.jetbrains.kotlin.js.descriptorUtils.getJetTypeFqName
-import org.jetbrains.kotlin.js.resolve.diagnostics.findPsi
 import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.nj2k.postProcessing.type
 import org.jetbrains.kotlin.psi.*
-import org.jetbrains.kotlin.psi.psiUtil.containingClass
-import org.jetbrains.kotlin.psi.psiUtil.referenceExpression
-import org.jetbrains.kotlin.resolve.bindingContextUtil.getReferenceTargets
-import org.jetbrains.kotlin.resolve.calls.callUtil.getType
-import org.jetbrains.kotlin.resolve.descriptorUtil.fqNameSafe
-import org.jetbrains.kotlin.resolve.lazy.BodyResolveMode
-import org.jetbrains.kotlin.types.typeUtil.*
 
 class RegisterSignalInspection : AbstractKotlinInspection() {
     private val mutabilityQuickFix by lazy { RegisterSignalMutabilityQuickFix() }
