+++ kt/plugins/godot-intellij-plugin/src/main/kotlin/godot/intellij/plugin/inspections/registration/quickfix/EngineFunctionNotRegisteredQuickFix.kt
@@ -1,4 +1,4 @@
-package godot.intellij.plugin.inspections.registration.property.quickfix
+package godot.intellij.plugin.inspections.registration.quickfix
 
 import com.intellij.codeInspection.LocalQuickFix
 import com.intellij.codeInspection.ProblemDescriptor
@@ -7,7 +7,6 @@ import godot.annotation.RegisterFunction
 import org.jetbrains.kotlin.idea.util.addAnnotation
 import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.psi.*
-import org.jetbrains.kotlin.resolve.ImportPath
 
 class EngineFunctionNotRegisteredQuickFix : LocalQuickFix {
     override fun getFamilyName(): String {
